move config file update from tunneler lib to ZET
tunneler SDK should not make assumptions on how config is stored
